backend/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
numpy
scikit-learn
requests
shap
# AI/ML
torch --index-url https://download.pytorch.org/whl/cpu
transformers
einops
sentencepiece
supabase
biopython
xgboost
umap-learn
hdbscan
